pyproject.toml
[project]
name = "qa-pairs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.32.0",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "lefthook>=1.12.2",
    "ruff>=0.12.5",
    "ty>=0.0.1a16",
]
